Add fully working hybrid project
An application and library to show how breakpoints can be hit in native
library projects using onnly Android Studio and Gradle.

This was tested on Android Studio 2.0 Preiew 4 and uses the latest
gradle-experimental-0.6.0-alpha3 which uses a different Gradle DSL
compared to the standard Gradle plugin.

This is a stripped project without and Android Studio files.

For breakpoints to work, the library's path of the target ABI ned to be
added to the hybrid apps hybrid debugger.